gen.go: allow to customise the URL when generating
"go run gen.go" now takes the URL from the command line.
That URL is persisted in the go:generate line produced in certifi.go, so
"go generate" will reuse that URL forever.

This allows to easily fork the project to use a different URL (such as
the original Mozilla) if someone doesn't trust mkcert.org.